pyproject.toml
[project]
name = "tr-shared-lib"
version = "0.9.6"
description = "Shared library for ThinkRealty microservices"
requires-python = ">=3.13"
dependencies = [
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
redis = ["redis>=5.0.0"]
celery = ["celery>=5.3.0", "cachetools>=5.0.0"]
db = [
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
]
http = ["httpx>=0.27.0"]
middleware = [
    "fastapi>=0.109.0",
    "httpx>=0.27.0",
]
logging = ["structlog>=24.0.0"]
cache = ["redis>=5.0.0"]
upstash = ["upstash-redis>=1.0.0"]
auth = [
    "pyjwt>=2.8.0",
    "cryptography>=42.0.0",
]
monitoring = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-prometheus>=0.44b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.44b0",
    "prometheus-client>=0.20.0",
    "fastapi>=0.109.0",
    "httpx>=0.27.0",
    "redis>=5.0.0",
]
# Provider-specific monitoring extras (for lean installations)
monitoring-prometheus = [
    "opentelemetry-exporter-prometheus>=0.44b0",
    "prometheus-client>=0.20.0",
]
monitoring-otlp = [
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
]
rate-limiter = [
    "redis>=5.0.0",
    "fastapi>=0.109.0",
]
webhooks = [
    "fastapi>=0.109.0",
    "redis>=5.0.0",
]
all = [
    "tr-shared-lib[redis,celery,db,http,middleware,logging,auth,monitoring,cache,rate-limiter,webhooks]",
]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tr_shared"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=src/tr_shared --cov-report=term-missing"

[tool.coverage.run]
source = ["src/tr_shared"]
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
fail_under = 80
show_missing = true
skip_covered = false

[dependency-groups]
dev = [
    "tr-shared-lib[all]",
    "pytest>=8.0.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "fakeredis>=2.34.1",
    "httpx>=0.28.1",
    "fastapi>=0.135.2",
    "lupa>=2.6",
]
